@@ -2,7 +2,7 @@
   "name": "homespace-frontend",
   "description": "",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "main.jsx",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
@@ -20,53 +20,45 @@
   },
   "homepage": "https://github.com/UggoPrince/homespace-frontend#readme",
   "dependencies": {
-    "@apollo/client": "^3.3.14",
+    "@apollo/client": "^3.6.9",
     "@emotion/react": "^11.8.2",
     "@emotion/styled": "^11.8.1",
     "@mui/icons-material": "^5.5.1",
-    "@mui/material": "^5.5.0",
+    "@mui/material": "^5.8.7",
     "apollo-link-http": "^1.5.17",
-    "bootstrap": "^5.1.3",
-    "dotenv": "^8.2.0",
-    "graphql": "^15.5.0",
-    "graphql-tag": "^2.11.0",
-    "node-sass": "^6.0.1",
+    "apollo-upload-client": "^17.0.0",
+    "country-to-currency": "^1.0.8",
+    "dotenv": "^16.0.1",
+    "graphql": "^15.8.0",
+    "graphql-tag": "^2.12.6",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-hook-form": "^7.27.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hot-toast": "^2.1.1",
     "react-icons": "^4.2.0",
-    "react-paginate": "^7.1.3",
-    "react-redux": "^7.2.4",
+    "react-paginate": "^8.1.3",
+    "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
-    "redux": "^4.1.0",
-    "regenerator-runtime": "^0.13.9",
-    "tailwindcss": "^2.2.16"
+    "redux": "^4.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.13.10",
+    "@babel/eslint-parser": "^7.18.2",
     "@babel/plugin-syntax-class-properties": "^7.12.13",
     "@babel/preset-env": "^7.13.10",
     "@babel/preset-react": "^7.12.13",
     "@vitejs/plugin-react": "^1.0.0",
-    "babel-eslint": "^10.1.0",
+    "autoprefixer": "^10.4.7",
     "babel-loader": "^8.2.2",
-    "css-loader": "^5.2.0",
     "eslint": "^7.22.0",
     "eslint-config-airbnb": "^18.2.1",
-    "eslint-loader": "^4.0.2",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.23.0",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "file-loader": "^6.2.0",
-    "postcss": "^8.2.9",
-    "postcss-loader": "^5.2.0",
-    "postcss-preset-env": "^6.7.0",
-    "react-hot-loader": "^4.13.0",
-    "sass-loader": "^12.2.0",
-    "style-loader": "^2.0.0",
+    "eslint-webpack-plugin": "^3.2.0",
+    "postcss": "^8.4.14",
+    "tailwindcss": "^3.1.6",
     "vite": "^2.6.4",
     "webpack": "^5.27.1",
     "webpack-cli": "^4.5.0",
