@@ -1,11 +1,11 @@
 /* eslint-disable import/no-extraneous-dependencies */
-const tailwindcss = require('tailwindcss');
+// const tailwindcss = require('tailwindcss');
 // const autoPrefixer = require('autoprefixer');
 
 module.exports = {
-  plugins: [
-    tailwindcss,
-    // autoPrefixer,
+  plugins: {
+    tailwindcss: {},
+    autoPrefixer: {},
     // 'postcss-preset-env',
-  ],
+  },
 };
