@@ -1,8 +1,9 @@
-import React from 'react';
+import React, { useRef } from 'react';
 import { connect } from 'react-redux';
 import {
   useNavigate,
 } from 'react-router-dom';
+import { useApolloClient } from '@apollo/client';
 import {
   destroyLocalStorage, getLocalStorage, getUserFromLocalStorage, setLocalStorage,
 } from '../Utils/LocalStorage';
@@ -14,13 +15,36 @@ const AuthProvider = (props) => {
   const { children, q, propsSearchOffset } = props;
   const start = propsSearchOffset;
   const navigate = useNavigate();
-  const [token, setToken] = React.useState(getLocalStorage('token'));
-  const [user, setUser] = React.useState(getUserFromLocalStorage('user'));
-  const handleLogin = (token, user) => {
-    setLocalStorage('user', JSON.stringify(user));
-    setLocalStorage('token', token);
-    setToken(token);
-    setUser(user);
+  const token = useRef(getLocalStorage('token'));
+  const user = useRef(getUserFromLocalStorage('user'));
+  const updateUser = () => {
+    user.current = getUserFromLocalStorage('user');
+  };
+  const isAuthenticated = useRef(token.current !== null && token.current !== undefined);
+  const updateIsAuth = (value) => {
+    isAuthenticated.current = value;
+  };
+  const path = window.location.pathname.toLowerCase();
+  const str = path === '/' ? 'home' : path.replace('/', '');
+  const isActive = useRef(str);
+  const isActiveMenuItem = useRef('');
+  const client = useApolloClient();
+
+  const setIsActive = (p) => {
+    isActive.current = p;
+  };
+
+  const setIsActiveMenuItem = (p) => {
+    isActiveMenuItem.current = p;
+  };
+
+  const handleLogin = (tokenStr, userObj) => {
+    setLocalStorage('user', JSON.stringify(userObj));
+    setLocalStorage('token', tokenStr);
+    token.current = tokenStr;
+    user.current = userObj;
+    isActive.current = 'home';
+    updateIsAuth(true);
     let url = '/';
     if (q && !start) url += computeSearchPath(q, start);
     else if (!q && start) url += computeSearchPath(q, start);
@@ -28,19 +52,34 @@ const AuthProvider = (props) => {
     navigate(url, { replace: true });
   };
 
-  const handleLogout = () => {
+  const clearStorage = () => {
     destroyLocalStorage('token');
     destroyLocalStorage('user');
-    setToken(null);
-    setUser(null);
-    if (!itsThisPath('/')) navigate('/');
+    token.current = null;
+    user.current = null;
+    updateIsAuth(false);
+  };
+
+  const handleLogout = () => {
+    clearStorage();
+    client.clearStore();
+    if (!itsThisPath('/')) navigate('/login', { replace: true });
+    else navigate('/');
   };
 
   const value = {
-    token,
-    user,
+    token: token.current,
+    user: user.current,
+    updateUser,
+    isAuthenticated: isAuthenticated.current,
+    updateIsAuth,
     loginUser: handleLogin,
     logout: handleLogout,
+    clearStorage,
+    isActive,
+    setIsActive,
+    isActiveMenuItem,
+    setIsActiveMenuItem,
   };
 
   return (
