@@ -1,20 +1,36 @@
 /* eslint-disable react/button-has-type */
+import './style.css';
+
 export default (props) => {
-  const { id, type, text } = props;
+  const {
+    id, type, text, handleclick,
+  } = props;
   return (
     <button
       id={id}
       type={type}
-      className=" w-full
-                bg-indigo-600
-                hover:bg-blue-700
-                focus:outline-none
-                focus:ring-2
-                focus:ring-purple-600
-                focus:ring-opacity-50
-                text-white
-                py-2 px-2
-                rounded font-bold"
+      onClick={handleclick}
+      className="default-button bg-indigo-600"
+    >
+      {text}
+    </button>
+  );
+};
+
+export const CustomButton = (props) => {
+  const {
+    id, type, text, handleclick, classname, inactive = false,
+  } = props;
+  return (
+    <button
+      disabled={inactive}
+      id={id}
+      type={type}
+      onClick={handleclick}
+      className={`${(
+        classname
+        || 'bg-indigo-600 hover:bg-blue-700 focus:ring-purple-600')}
+       custom-button`}
     >
       {text}
     </button>
