@@ -1,6 +1,11 @@
-import React, { useState } from 'react';
+import { useState, useRef } from 'react';
 
 export default () => {
   const [errors, setErrors] = useState({});
   return { errors, setErrors };
+};
+
+export const useRefForm = () => {
+  const errors = useRef({});
+  return errors;
 };
