@@ -1,5 +1,7 @@
 import React, { useState } from 'react';
-import { TextField, InputAdornment, IconButton } from '@mui/material';
+import {
+  TextField, InputAdornment, IconButton, MenuItem,
+} from '@mui/material';
 import Visibility from '@mui/icons-material/Visibility';
 import VisibilityOff from '@mui/icons-material/VisibilityOff';
 
@@ -61,5 +63,84 @@ export const PasswordTextField = (props) => {
   </InputAdornment>,
       }}
     />
+  );
+};
+
+export const TextArea = (props) => {
+  const {
+    name, label, error, onchange, type, required,
+  } = props;
+  return (
+    <TextField
+      name={name}
+      label={label}
+      className="w-full"
+      multiline
+      maxRows={2}
+      minRows={2}
+      onChange={onchange}
+      type={type}
+      required={required}
+      {...(error && { error: true, helperText: error })}
+    />
+  );
+};
+
+export const SelectField = (props) => {
+  const [value, setValue] = useState('');
+  let index = -1;
+  const {
+    options = [''], label, error = null, name, extraFunction,
+  } = props;
+  const handleChange = (event) => {
+    setValue(event.target.value);
+    if (extraFunction) {
+      extraFunction(event.target.value);
+    }
+  };
+  return (
+    <TextField
+      name={name}
+      variant="outlined"
+      className="w-full max-h-20"
+      value={value}
+      select
+      SelectProps={{
+        MenuProps: {
+          anchorOrigin: {
+            vertical: 'bottom',
+            horizontal: 'center',
+          },
+          transformOrigin: {
+            vertical: 'top',
+            horizontal: 'center',
+          },
+          PaperProps: { sx: { maxHeight: 200 } },
+        },
+      }}
+      label={label}
+      {...(error && { error: true, helperText: error })}
+      onChange={handleChange}
+    >
+      {options.length > 1 && options.map((item, i) => (
+        <MenuItem
+          // eslint-disable-next-line react/no-array-index-key
+          key={item}
+          value={item}
+          className="w-full m-auto"
+          onClick={(event) => { index = i; }}
+        >
+          {item}
+        </MenuItem>
+      ))}
+      {options.length < 1
+      && (
+      <MenuItem
+        value=""
+      >
+        {label}
+      </MenuItem>
+      )}
+    </TextField>
   );
 };
