@@ -1,6 +1,51 @@
 const varToString = (varObj) => Object.keys(varObj)[0];
 
-const handleError = (err, divId) => {
+// extract variables from login form
+export const getLoginVariables = (event) => {
+  const {
+    email, password,
+  } = event.target;
+  return {
+    email: email.value,
+    password: password.value,
+  };
+};
+
+// extract variables from signup form
+export const getSignupVariables = (event) => {
+  const {
+    firstname, lastname, email, password, confirmPassword, country, state, address,
+  } = event.target;
+  return {
+    firstname: firstname.value,
+    lastname: lastname.value,
+    email: email.value,
+    password: password.value,
+    country: country.value,
+    state: state.value,
+    address: address.value,
+  };
+};
+
+// login function
+export const signIn = (event, func) => {
+  event.preventDefault();
+  const variables = getLoginVariables(event);
+  func({
+    variables,
+  });
+};
+
+// signup function
+export const register = (event, func) => {
+  event.preventDefault();
+  const variables = getSignupVariables(event);
+  func({
+    variables,
+  });
+};
+
+/* const handleError = (err, divId) => {
   const div = document.getElementById(`${divId}`);
   // eslint-disable-next-line no-console
   console.log(div);
@@ -63,4 +108,4 @@ export const handleLoginErrors = (error, postfix) => {
     const divId = varToString({ password }) + postfix;
     handleError(password, divId);
   }
-};
+}; */
