@@ -2,7 +2,6 @@ import { Component } from 'react';
 import Header from '../../components/Header';
 import Footer from '../../components/Footer';
 import SignupForm from './signupForm';
-// import { prepareSignup } from '../../utils/EventHandlers';
 
 class Signup extends Component {
   componentDidMount() {
@@ -10,7 +9,7 @@ class Signup extends Component {
 
   render() {
     return (
-      <div className="2xl:container min-ht-vh relative">
+      <div className="w-full min-ht-vh relative master">
         <Header />
         <SignupForm />
         <Footer />
